refactor:커뮤니티리팩토링 및 커뮤니티 카드수정
