+++ frontend/src/app/community/[id]/write/page.tsx
@@ -1,11 +1,10 @@
 'use client';
 
-import React, { useState, use } from 'react';
-import { useRouter } from 'next/navigation';
+import React from 'react';
 import { InputField, Dropdown, Button } from '@/components/ui';
-import { useUIStore } from '@/store/useUIStore';
 import dynamic from 'next/dynamic';
 import styles from './page.module.css';
+import { useCommunityWrite } from './useCommunityWrite';
 
 const TiptapEditor = dynamic(() => import('@/components/editor/TiptapEditor'), { ssr: false });
 
@@ -14,55 +13,17 @@ interface Props {
 }
 
 export default function CommunityWritePage({ params }: Props) {
-  const router = useRouter();
-  const { showToast } = useUIStore();
-  
-  const resolvedParams = use(params);
-  const communityId = resolvedParams.id;
-
-  const [title, setTitle] = useState('');
-  const [content, setContent] = useState('');
-  const [type, setType] = useState('GENERAL');
-  const [isSubmitting, setIsSubmitting] = useState(false);
-
-  const handleSubmit = async () => {
-    if (!title.trim() || !content.trim() || content === '<p></p>') {
-      showToast('입력 오류', 'error', '제목과 내용을 모두 입력해주세요.');
-      return;
-    }
-
-    setIsSubmitting(true);
-
-    try {
-      const payload = {
-        title,
-        content,
-        type,
-        communityId: Number(communityId),
-      };
-
-      const response = await fetch('/api/posts', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify(payload),
-      });
-
-      if (!response.ok) {
-        throw new Error('게시글 등록에 실패했습니다.');
-      }
-
-      showToast('게시글 등록 완료', 'success', '성공적으로 글을 작성했습니다.');
-      
-      router.push(`/community/${communityId}`);
-    } catch (error) {
-      console.error(error);
-      showToast('등록 실패', 'error', '서버 오류가 발생했습니다.');
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
+  const {
+    title,
+    setTitle,
+    content,
+    setContent,
+    type,
+    setType,
+    isSubmitting,
+    handleSubmit,
+    router
+  } = useCommunityWrite({ params });
 
   return (
     <div className={styles.container}>
