+++ frontend/src/app/community/_components/CommunityCard/CommunityCard.tsx
