+++ frontend/src/app/community/_components/CommunityCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CommunityCard';
