+++ frontend/src/app/community/_components/CommunityCommentItem/CommunityCommentItem.tsx
