+++ frontend/src/app/community/_components/CommunityCommentItem/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CommunityCommentItem';
