+++ frontend/src/app/community/_components/CommunityDetailHeader/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CommunityDetailHeader';
