+++ frontend/src/app/community/_components/CommunityPostContainer/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CommunityPostContainer';
