+++ frontend/src/app/community/_components/CommunityPostItem/CommunityPostItem.tsx
