+++ frontend/src/app/community/_components/CommunityPostItem/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CommunityPostItem';
